requirements.txt
python-jobspy>=1.1.70
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.12.0
openpyxl>=3.1.0
pandas>=2.2.0
streamlit>=1.32.0
python-dotenv>=1.0.0
lxml>=5.1.0
